fix(storage/ipc): catalog horizon for logical-decoding rels, wrapped-xid recovery sort, checkpoint-delay snapshot, pg_terminate_backend timeout, dsm/ipc conformance (b020) (#1666)
* test(storage/ipc): witnesses for the b020 audit rows (fail on main)

Regression witnesses committed ahead of their fixes, all asserting the C 18.6
outcome:

- procarray: ProcArrayApplyRecoveryInfo must order a wrap-straddling
  running-xacts snapshot modularly (procarray.c:1204 xidLogicalComparator);
  the checkpoint delay wait must key on the snapshotted vxids, not on
  newcomers (procarray.c:3043/3084 + xlog.c CreateCheckPoint); a
  ProcArrayGroupClearXid follower must report ProcarrayGroupUpdate
  (procarray.c:827/836).
- dsm_core: an erroring on-detach callback must not leak
  InterruptHoldoffCount (dsm.c:813 HOLD/RESUME_INTERRUPTS); an unsizeable
  create request is an ereport, not a panic (dsm_impl.c:366); allocation
  reports DsmAllocate (dsm_impl.c:366).
- signalfuncs: pg_terminate_backend narrows its int64 timeout to int before
  the sign check (signalfuncs.c:241).
- ipc: cancel_before_shmem_exit prints the full Datum word (ipc.c:409);
  check_on_shmem_exit_lists_are_empty is FATAL (ipc.c:442/444).
- scripts/ipc-conformance-e2e.sh: both engines under wal_level=logical —
  pg_terminate_backend timeout narrowing and the user_catalog_table catalog
  horizon (procarray.c:1985 RelationIsAccessibleInLogicalDecoding) byte-
  compared against C 18.6.


* fix(procarray): logical-decoding relations use the catalog horizon

procarray.c:1985 GlobalVisHorizonKindForRel returns VISHORIZON_CATALOG for
IsCatalogRelation(rel) || RelationIsAccessibleInLogicalDecoding(rel). The
second arm (wal_level=logical, RelationNeedsWAL, user_catalog_table=true) was
dropped as unreachable when wal_level=logical was unported; with logical
slots shipped, VACUUM on a user_catalog_table pruned tuples a slot's
catalog_xmin still needed (C keeps them: "50 are dead but not yet removable").

Witness: scripts/ipc-conformance-e2e.sh (VACUUM VERBOSE under a logical slot,
byte-compared with C 18.6).


* fix(procarray,xlog): checkpoint delay waits for the snapshotted vxids only

procarray.c:3043 GetVirtualXIDsDelayingChkpt snapshots the vxids holding the
delay flag and procarray.c:3084 HaveVirtualXIDsDelayingChkpt rechecks only
those; xlog.c CreateCheckPoint takes the snapshot once and loops (do-while,
10 ms) until those transactions have left their critical sections. The
folded probe ("any current holder") made every transaction that entered a
critical section during the wait extend it, so a continuous commit stream
could starve the checkpointer. Restores the C shape: two seams, the
snapshot in wait_for_delay_chkpt, C's do-while cadence.

Witness: procarray::tests::delaying_chkpt_wait_tracks_the_snapshot_not_newcomers
(the same scenario as the witness commit, called through the C signatures).


* fix(procarray): report ProcarrayGroupUpdate while a group-clear follower sleeps

procarray.c:827/836 brackets the follower's semaphore sleep in
ProcArrayGroupClearXid with pgstat_report_wait_start(WAIT_EVENT_PROCARRAY_GROUP_UPDATE)
/ pgstat_report_wait_end, so pg_stat_activity shows wait_event
ProcarrayGroupUpdate for committing followers. waitevent_seams becomes a
production dependency of procarray (it was dev-only).

Witness: procarray::tests::group_clear_xid_follower_reports_procarray_group_update_wait.


* fix(procarray): sort a recovery running-xacts snapshot modularly

procarray.c:1204 sorts the incoming xids with xidLogicalComparator (modular
TransactionIdPrecedes order). A snapshot that straddles the 2^32 wraparound
(4294967294, 4294967295 and 3 all running) must insert the pre-wrap xids
first; plain unsigned order put 3 first and KnownAssignedXidsAdd then failed
recovery with "out-of-order XID insertion in KnownAssignedXids".

Witness: procarray::tests::apply_recovery_info_orders_wrapped_xids_modularly.


* fix(dsm): release the interrupt hold when a detach callback errors

dsm.c:813 runs the on-detach callbacks between HOLD_INTERRUPTS() and
RESUME_INTERRUPTS(); a callback ERROR longjmps to a handler that resets
InterruptHoldoffCount. The Rust ? early-return skipped ResumeInterrupts, so
an erroring callback left the holdoff count raised — on the Drop path (error
demoted to WARNING) for the rest of the session, disabling query cancel and
statement_timeout. Run the callbacks in a closure and resume on both paths;
the dsm test that compensated with a trailing ResumeInterrupts drops it.

Witness: dsm_core::tests::erroring_detach_callback_does_not_leak_interrupt_holdoff.


* fix(dsm): report an unsizeable segment request instead of panicking

dsm_impl.c:366 dsm_impl_posix_resize: when ftruncate/posix_fallocate cannot
size the segment (EFBIG/ENOMEM), the failure is an ereport at the caller's
elevel — "could not resize shared memory segment \"%s\" to %zu bytes: %m"
with errcode_for_dynamic_shared_memory() = ERRCODE_OUT_OF_MEMORY — and
DSM_OP_CREATE returns false. The heap-region model called
Layout::from_size_align(..).expect(), so a request above isize::MAX aborted
the thread with "dsm region layout". region_layout now yields None and the
create arm reports EFBIG at elevel.

Witness: dsm_core::tests::create_reports_unrepresentable_size_as_error_not_panic.


* fix(dsm): report DsmAllocate while a segment is allocated

dsm_impl.c:366 brackets the segment allocation with
pgstat_report_wait_start(WAIT_EVENT_DSM_ALLOCATE) / pgstat_report_wait_end,
so pg_stat_activity shows wait_event DsmAllocate (class IO) during DSM
creation. waitevent_seams becomes a production dependency of dsm_core (it
was dev-only).

Witness: dsm_core::tests::create_reports_dsm_allocate_wait_event.


* fix(signalfuncs): pg_terminate_backend narrows its timeout to int like C

signalfuncs.c:241 declares `int timeout` and assigns PG_GETARG_INT64(1) to
it before the sign check, so 3000000000 wraps negative and raises 22003
"timeout" must not be negative, and 4294967296 wraps to 0 (no wait). pgrust
kept the i64 and would instead wait up to 3e9 ms.

Witness: signalfuncs::tests::terminate_timeout_narrows_to_int_like_c and
scripts/ipc-conformance-e2e.sh (byte-compared with C 18.6).


* fix(ipc): cancel_before_shmem_exit prints the whole Datum word

ipc.c:409 formats the mismatched entry's arg as 0x%PRIxPTR (the full
pointer-sized Datum); pgrust printed arg.as_i32() as usize, truncating a
64-bit value to its low 32 bits (0x100000005 -> 0x5) and sign-extending
values with bit 31 set.

Witness: ipc::tests::cancel_before_shmem_exit_reports_the_full_datum_word.


* fix(ipc): check_on_shmem_exit_lists_are_empty is FATAL

ipc.c:442/444 raise elog(FATAL, ...) when a shmem-exit callback was
registered prematurely; pgrust returned an ERROR-level PgError, which a
caller could catch as an ordinary transaction error.

Witness: ipc::tests::check_lists_not_empty_is_fatal.


* fix(procarray): repair the mis-applied hunks of the three lib.rs fix commits

The per-root-cause split of lib.rs (95f07cf783b, 0937d9df92f, 3f93021e2f2)
was staged from zero-context hunks and landed two of them at the wrong
offsets (helper inserted inside GetCurrentVirtualXIDs, snapshot loop body
displaced), which does not compile (fleet job
pgrust-fast-tests-a4dac22e0d-1788577622-1fe9). This restores the intended
file: the same three changes, in place. No behavioural change beyond those
commits' descriptions.


* docs(audit): lane note + LANES row for b020-backend-storage-ipc-1


* fixup: sync Cargo.lock for dsm_core's waitevent_seams dependency

The dsm fix (4f4d0bae00f) added waitevent_seams to dsm_core's Cargo.toml
without the matching Cargo.lock entry, so the --locked lint gate refused
to resolve. procarray's promotion of waitevent_seams from dev-dependency
to dependency needs no lock change (already listed).


* docs: drop the per-lane notes/LANES.md row — the campaign index lands separately (queue conflict fix)

---------

+++ crates/backend/storage/ipc/signalfuncs/src/lib.rs
@@ -154,7 +154,10 @@ pub fn fc_pg_terminate_backend(
     fcinfo: &mut Fcinfo,
 ) -> PgResult<Datum> {
     let pid = fcinfo.arg_i32(0);
-    let timeout = fcinfo.arg_i64(1);
+    // signalfuncs.c:241 `int timeout = PG_GETARG_INT64(1)`: the int64 is
+    // narrowed to a C int before the sign check (3000000000 wraps negative,
+    // 4294967296 wraps to 0 = no wait).
+    let timeout = fcinfo.arg_i64(1) as i32;
 
     if timeout < 0 {
         return Err(elog::ereport(types_error::ERROR)
@@ -185,7 +188,7 @@ pub fn fc_pg_terminate_backend(
     }
 
     if r == SIGNAL_BACKEND_SUCCESS && timeout > 0 {
-        Ok(Datum::from_bool(pg_wait_until_termination(pid, timeout)?))
+        Ok(Datum::from_bool(pg_wait_until_termination(pid, i64::from(timeout))?))
     } else {
         Ok(Datum::from_bool(r == SIGNAL_BACKEND_SUCCESS))
     }
